+++ packages/dash-table/tests/cypress/tests/standalone/filtering_test.ts
@@ -74,19 +74,19 @@ describe('filter', () => {
         DashTable.getFilterById('ccc').click();
         DOM.focused.type(`gt`);
         DashTable.getFilterById('ddd').click();
-        DOM.focused.type('20 a000');
+        DOM.focused.type('"20 a000');
         DashTable.getFilterById('eee').click();
         DOM.focused.type('is prime2');
         DashTable.getFilterById('bbb').click();
-        DOM.focused.type('! !');
+        DOM.focused.type('! !"');
         DashTable.getFilterById('ccc').click();
 
         DashTable.getCellById(0, 'ccc').within(() => cy.get('.dash-cell-value').should('have.html', cell_0));
         DashTable.getCellById(1, 'ccc').within(() => cy.get('.dash-cell-value').should('have.html', cell_1));
 
-        DashTable.getFilterById('bbb').within(() => cy.get('input').should('have.value', '! !'));
+        DashTable.getFilterById('bbb').within(() => cy.get('input').should('have.value', '! !"'));
         DashTable.getFilterById('ccc').within(() => cy.get('input').should('have.value', 'gt'));
-        DashTable.getFilterById('ddd').within(() => cy.get('input').should('have.value', '20 a000'));
+        DashTable.getFilterById('ddd').within(() => cy.get('input').should('have.value', '"20 a000'));
         DashTable.getFilterById('eee').within(() => cy.get('input').should('have.value', 'is prime2'));
 
         DashTable.getFilterById('bbb').should('have.class', 'invalid');